==> ship_project/03_DB연동.ipynb <==
import pandas as pd
import mariadb
from sqlalchemy import create_engine

conn  = mariadb.connect(
    user='team3', 
    password='dbdb', 
    database='team3db', 
    host='localhost',
    port=3306
)

# 위 커넥션 정보와 동일하게 입력
engine = create_engine("mysql://{user}:{pw}@{host}/{db}".format(user='team3', pw='dbdb',host='localhost', db='team3db'))

df = pd.read_csv('./data/train.csv')

df.to_sql('train2', engine, if_exists='replace', chunksize=1000)

# SELECT문

select_sql = "SELECT * FROM train"

selected_df = pd.read_sql_query(select_sql, engine,index_col='index')

selected_df

df2 = pd.read_csv('./data/test.csv')
df2.to_sql('test',engine, if_exists='replace', chunksize=1000 )

selected_df = pd.read_sql_query('SELECT * FROM test', engine,index_col='index')
selected_df
