==> ship_project/05_db추가.ipynb <==
import pandas as pd
import mariadb
from sqlalchemy import create_engine
import matplotlib.pyplot as plt
from sklearn.preprocessing import LabelEncoder, MinMaxScaler
from sklearn.model_selection import train_test_split
from supervised.automl import AutoML
from tqdm import tqdm
import bisect
import numpy as np


conn  = mariadb.connect(
    user='team3', 
    password='dbdb', 
    database='team3db', 
    host='localhost',
    port=3306
)

# 위 커넥션 정보와 동일하게 입력
engine = create_engine("mysql://{user}:{pw}@{host}/{db}".format(user='team3', pw='dbdb',host='localhost', db='team3db'))

select_sql = "SELECT * FROM train2"
train = pd.read_sql_query(select_sql, engine,index_col='index')
train.reset_index(inplace=True,drop=True)
train

train = train.loc[train['ARI_PO']!='RGT8']

train.loc[train['ARI_PO']=='RGT8']

train.drop(columns=['SAMPLE_ID','ARI_PO','V_WIND'],inplace=True)

train

train['ATA'] = pd.to_datetime(train['ATA'])

train['year'] = train['ATA'].dt.year
train['weekday'] = train['ATA'].dt.weekday

train.drop(columns='ATA', inplace=True)
train.shape

train = train.loc[train['year']!=2014]
train.shape

train = train.loc[train['year']!=2023]
train.shape

train.dropna(inplace=True)
train.shape

train.drop(columns=['ID', 'SHIPMANAGER', 'BN', 'ATA_LT'], inplace=True)
train.shape

train.reset_index(drop=True)

q1, q3 = np.percentile(train['BUILT'],[25,75])
q1, q3

upper_bound = q3 + ((q3-q1) * 1.5)
upper_bound

train.loc[train['BUILT']>30, 'BUILT'] = upper_bound

train.to_sql('train3',engine,if_exists='replace',chunksize=1000)
