==> ship_project/07_linear_train_test_split.ipynb <==
import pandas as pd
import mariadb
from sqlalchemy import create_engine
import matplotlib.pyplot as plt
from sklearn.preprocessing import LabelEncoder, MinMaxScaler, StandardScaler
from supervised.automl import AutoML
from tqdm import tqdm
import numpy as np
import pandas as pd
import numpy as np
import bisect
from tqdm import tqdm
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score


conn  = mariadb.connect(
    user='team3', 
    password='dbdb', 
    database='team3db', 
    host='localhost',
    port=3306
)

# 위 커넥션 정보와 동일하게 입력
engine = create_engine("mysql://{user}:{pw}@{host}/{db}".format(user='team3', pw='dbdb',host='localhost', db='team3db'))

train = pd.read_sql_query('SELECT * FROM train_UL',engine,index_col='index')
train.reset_index(inplace=True,drop=True)
train.drop(columns=['SAMPLE_ID','ID','ARI_PO','V_WIND'],inplace=True)
train

test = pd.read_sql_query('SELECT * FROM test',engine, index_col='index')
test.reset_index(drop=True,inplace=True)
test

test.isnull().sum()

test.columns

test_port = test.groupby('ARI_PO')['U_WIND'].mean().to_dict()

for port, mean_temp in test_port.items():
    test.loc[test['ARI_PO'] == port, 'U_WIND'] = test.loc[test['ARI_PO'] == port, 'U_WIND'].fillna(mean_temp)
    

test['U_WIND'].isnull().sum()


test_port = test.groupby('ARI_PO')['AIR_TEMPERATURE'].mean().to_dict()

for port, mean_temp in test_port.items():
    test.loc[test['ARI_PO'] == port, 'AIR_TEMPERATURE'] = test.loc[test['ARI_PO'] == port, 'AIR_TEMPERATURE'].fillna(mean_temp)
    

test['AIR_TEMPERATURE'].isnull().sum()


test_port = test.groupby('ARI_PO')['BN'].mean().to_dict()

for port, mean_temp in test_port.items():
    test.loc[test['ARI_PO'] == port, 'BN'] = test.loc[test['ARI_PO'] == port, 'BN'].fillna(mean_temp)
    

test['BN'].isnull().sum()

test.drop(columns=['SAMPLE_ID','ID','ARI_PO','V_WIND'],inplace=True)

test.shape

test.isnull().sum()

categorical_features = ['ARI_CO', 'SHIP_TYPE_CATEGORY',  'FLAG']
encoders = {}

for feature in tqdm(categorical_features, desc="Encoding features"):
    le = LabelEncoder()
    train[feature] = le.fit_transform(train[feature].astype(str))
    le_classes_set = set(le.classes_)
    test[feature] = test[feature].map(lambda s: '-1' if s not in le_classes_set else s)
    le_classes = le.classes_.tolist()
    bisect.insort_left(le_classes, '-1')
    le.classes_ = np.array(le_classes)
    test[feature] = le.transform(test[feature].astype(str))
    encoders[feature] = le

train['ATA'] = pd.to_datetime(train['ATA'])
test['ATA'] = pd.to_datetime(test['ATA'])
train['year'] = train['ATA'].dt.year
train['weekday'] = train['ATA'].dt.weekday
test['year'] = test['ATA'].dt.year
test['weekday'] = test['ATA'].dt.weekday
train.drop(columns=['ATA','ATA_LT','BN','SHIPMANAGER'], inplace=True)
test.drop(columns=['ATA','ATA_LT','BN','SHIPMANAGER'], inplace=True)

test.fillna(test.mean(),inplace=True)

X_train = train['DIST'].to_numpy().reshape(-1,1)
Y_train = train['CI_HOUR']
test = test['DIST'].to_numpy().reshape(-1,1)

from sklearn.model_selection import train_test_split

train_input, test_input, train_target, test_target = train_test_split(X_train,Y_train, random_state=42, test_size=0.2)

from sklearn.linear_model import LinearRegression,LogisticRegression, Lasso, Ridge

lr = LinearRegression()
logistic = LogisticRegression(random_state=42)
lasso = Lasso(random_state=42)
ridge = Ridge(random_state=42)
lr.fit(train_input,train_target)
# logistic.fit(train_input,train_target)
lasso.fit(train_input,train_target)
ridge.fit(train_input,train_target)

pred_lr = lr.predict(test_input)
pred_logistic = logistic.predict(test_input)
pred_lasso = lasso.predict(test_input)
pred_ridge = ridge.predict(test_input)
mae_lr = mean_absolute_error(test_input,pred_lr)
mse_lr = mean_squared_error(test_input,pred_lr)
r2_lr = r2_score(test_input, pred_lr)

print(mae_lr, mse_lr, r2_lr)

# mae_logistic = mean_absolute_error(test_input,pred_logistic)
# mse_logistic = mean_squared_error(test_input,pred_logistic)
# r2_logistic = r2_score(test_input, pred_logistic)
# print(mae_logistic, mse_logistic, r2_logistic)

mae_lasso = mean_absolute_error(test_input,pred_lasso)
mse_lasso = mean_squared_error(test_input,pred_lasso)
r2_lasso = r2_score(test_input, pred_lasso)
print(mae_lasso, mse_lasso, r2_lasso)

mae_lasso = mean_absolute_error(test_input,pred_lasso)
mse_lasso = mean_squared_error(test_input,pred_lasso)
r2_lasso = r2_score(test_input, pred_lasso)
print(mae_lasso, mse_lasso, r2_lasso)

pred_lr2 = lr.predict(test)
pred_lasso2 = lasso.predict(test)
# pred_logistic2 = logistic.predict(test)
pred_ridge2 = ridge.predict(test)

submit = pd.read_csv('./data/sample_submission.csv')
submit['CI_HOUR'] = pred_lr2
submit.to_csv('./csv/submit8.csv', index=False)

# submit = pd.read_csv('./data/sample_submission.csv')
# submit['CI_HOUR'] = pred_logistic2
# submit.to_csv('./csv/submit9.csv', index=False)

submit = pd.read_csv('./data/sample_submission.csv')
submit['CI_HOUR'] = pred_ridge2
submit.to_csv('./csv/submit10.csv', index=False)

submit = pd.read_csv('./data/sample_submission.csv')
submit['CI_HOUR'] = pred_lasso2
submit.to_csv('./csv/submit11.csv', index=False)
