==> ship_project/08_linear_model.ipynb <==
import pandas as pd
import mariadb
from sqlalchemy import create_engine
import matplotlib.pyplot as plt
from sklearn.preprocessing import LabelEncoder, MinMaxScaler, StandardScaler
from sklearn.linear_model import LinearRegression,LogisticRegression, Lasso, Ridge
import numpy as np
import bisect
from tqdm import tqdm
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score


conn  = mariadb.connect(
    user='team3', 
    password='dbdb', 
    database='team3db', 
    host='localhost',
    port=3306
)

# 위 커넥션 정보와 동일하게 입력
engine = create_engine("mysql://{user}:{pw}@{host}/{db}".format(user='team3', pw='dbdb',host='localhost', db='team3db'))

train = pd.read_sql_query('SELECT * FROM train_UL',engine,index_col='index')

test = pd.read_sql_query('SELECT * FROM test',engine, index_col='index')


X_train = train['DIST'].to_numpy().reshape(-1,1)
Y_train = train['CI_HOUR']
test = test['DIST'].to_numpy().reshape(-1,1)

lr = LinearRegression()
lasso = Lasso(random_state=42)
ridge = Ridge(random_state=42)
lr.fit(X_train,Y_train)
lasso.fit(X_train,Y_train)
ridge.fit(X_train,Y_train)

pred_lr = lr.predict(test)
pred_lasso = lasso.predict(test)
pred_ridge = ridge.predict(test)
mae_lr = mean_absolute_error(test,pred_lr)
mse_lr = mean_squared_error(test,pred_lr)
r2_lr = r2_score(test, pred_lr)

print(mae_lr, mse_lr, r2_lr)

mae_lasso = mean_absolute_error(test,pred_lasso)
mse_lasso = mean_squared_error(test,pred_lasso)
r2_lasso = r2_score(test, pred_lasso)
print(mae_lasso, mse_lasso, r2_lasso)

mae_ridge = mean_absolute_error(test,pred_ridge)
mse_ridge = mean_squared_error(test,pred_ridge)
r2_ridge = r2_score(test, pred_ridge)
print(mae_ridge, mse_ridge, r2_ridge)

submit = pd.read_csv('./data/sample_submission.csv')
submit['CI_HOUR'] = pred_lr
submit.to_csv('./csv/submit12.csv', index=False)

submit = pd.read_csv('./data/sample_submission.csv')
submit['CI_HOUR'] = pred_ridge
submit.to_csv('./csv/submit13.csv', index=False)

submit = pd.read_csv('./data/sample_submission.csv')
submit['CI_HOUR'] = pred_lasso
submit.to_csv('./csv/submit14.csv', index=False)
