==> ship_project/11_feature_engineering.ipynb <==
import pandas as pd
import mariadb
from sqlalchemy import create_engine
import matplotlib.pyplot as plt
from sklearn.preprocessing import LabelEncoder, MinMaxScaler, StandardScaler
from sklearn.ensemble import *
import numpy as np
import bisect
from tqdm import tqdm
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import xgboost as xgb
import lightgbm as lgb
import statsmodels.api as sm
from statsmodels.formula.api import ols
import seaborn as sns

conn  = mariadb.connect(
    user='team3', 
    password='dbdb', 
    database='team3db', 
    host='localhost',
    port=3306
)

# 위 커넥션 정보와 동일하게 입력
engine = create_engine("mysql://{user}:{pw}@{host}/{db}".format(user='team3', pw='dbdb',host='localhost', db='team3db'))

select_sql = "SELECT * FROM train_UL"
df = pd.read_sql_query(select_sql, engine,index_col='index')
df.reset_index(inplace=True,drop=True)
df.drop(columns='SAMPLE_ID',inplace=True)
df

df.isnull().sum()

df_nation = df.groupby('ARI_CO')['CI_HOUR'].std().to_dict()
df_port = df.groupby('ARI_PO')['CI_HOUR'].std()
df_ship = df.groupby('SHIP_TYPE_CATEGORY')['CI_HOUR'].std()
df_flag = df.groupby('FLAG')['CI_HOUR'].std()

df['FLAG'].isnull().sum()

for port, mean_temp in df_nation.items():
    df.loc[df['ARI_CO']==port, 'ARI_CO'] = mean_temp
    

df=df.astype({'ARI_CO':'float'})

df['ARI_CO']

for port, mean_temp in df_port.items():
    df.loc[df['ARI_PO']==port, 'ARI_PO'] = mean_temp
df=df.astype({'ARI_PO':'float'})

for port, mean_temp in df_ship.items():
    df.loc[df['SHIP_TYPE_CATEGORY']==port, 'SHIP_TYPE_CATEGORY'] = mean_temp
df=df.astype({'SHIP_TYPE_CATEGORY':'float'})

for port, mean_temp in df_flag.items():
    df.loc[df['FLAG']==port, 'FLAG'] = mean_temp
df=df.astype({'FLAG':'float'})

df['FLAG'].isnull().sum()

df.info()

df.isnull().sum()

df['ATA'] = pd.to_datetime(df['ATA'])
df['year'] =df['ATA'].dt.year
df['month'] =df['ATA'].dt.month
df['day'] =df['ATA'].dt.day
df['hour'] =df['ATA'].dt.hour
df['minute'] =df['ATA'].dt.minute
df['weekday'] =df['ATA'].dt.weekday

df.drop(columns='ATA',inplace=True)

df['FLAG'].fillna(df['FLAG'].mean(),inplace=True)

X = df[['ARI_CO', 'ARI_PO', 'SHIP_TYPE_CATEGORY', 'DIST', 'BREADTH', 'BUILT', 'DEADWEIGHT', 'DEPTH', 'DRAUGHT', 'GT', 'LENGTH', 
            'FLAG', 'U_WIND', 'V_WIND', 'AIR_TEMPERATURE', 'PORT_SIZE', 'year', 'weekday','BN', 'ATA_LT', 'month', 'day', 'hour', 'minute']]
y = df['CI_HOUR']

X.info()

import statsmodels.api as sm

# 독립 변수에 상수항 추가
X_with_const = sm.add_constant(X)

# 회귀 모델 생성
model = sm.OLS(y, X_with_const)

# 모델 피팅
results = model.fit()

# 결과 요약
print(results.summary())

from statsmodels.stats.outliers_influence import variance_inflation_factor

df_vif = pd.DataFrame()
df_vif['VIF'] = np.round([variance_inflation_factor(X.values,i) for i in range(X.shape[1])],3)
df_vif['feature'] = X.columns
df_vif.sort_values(by='VIF',ascending=False)

df.boxplot(column=['DRAUGHT'])

# df.to_csv('./data/train_std.csv')

X = df[['ARI_CO', 'ARI_PO', 'SHIP_TYPE_CATEGORY', 'DIST', 'BREADTH', 'BUILT', 'DEADWEIGHT', 'DEPTH', 'DRAUGHT', 'GT', 'LENGTH', 
            'FLAG',  'V_WIND', 'AIR_TEMPERATURE', 'PORT_SIZE', 'year', 'weekday', 'ATA_LT', 'month']]
y = df['CI_HOUR']

# 독립 변수에 상수항 추가
X_with_const = sm.add_constant(X)

# 회귀 모델 생성
model = sm.OLS(y, X_with_const)

# 모델 피팅
results = model.fit()

# 결과 요약
print(results.summary())

from statsmodels.stats.outliers_influence import variance_inflation_factor

df_vif = pd.DataFrame()
df_vif['VIF'] = np.round([variance_inflation_factor(X.values,i) for i in range(X.shape[1])],3)
df_vif['feature'] = X.columns
df_vif.sort_values(by='VIF',ascending=False)
